pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.15",
    "fastapi>=0.116.1",
    "fastmcp>=2.10.6",
    "flask>=3.1.2",
    "matplotlib>=3.10.5",
    "numpy>=2.3.2",
    "pandas>=2.3.2",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.2",
    "requests>=2.32.5",
    "scipy>=1.16.1",
    "seaborn>=0.13.2",
    "uvicorn>=0.35.0",
    "werkzeug>=3.1.3",
]
